training-platform core: extract training codegen helpers to trainingUtils.py
generateTrainingNetwork.py and testMVPTraining.py carried byte-identical
copies of _load_reference_losses, _infer_num_data_inputs, _infer_total_mb,
_infer_data_size, _infer_n_accum and the _GRAD_ACC constant.
testMVPOptimizer.py and testMVPTraining.py each defined their own
_mockScheduler, also byte-identical.

Move all of these to a new testUtils/trainingUtils.py module and import
them from the three entry points so the next helper edit only has to
happen in one place.  No behaviour change; the three entry-point scripts
still exist as before (user asked to keep testMVP*.py and the two
deeployTrainingRunner_siracusa.py / _tiled_siracusa.py stubs).

Verified on Siracusa: simplemlp_train passes 0/4 (diff=0.000000 at every
step) in both non-tiled and tiled runs.
