+++ DeeployTest/testUtils/trainingUtils.py
@@ -1,0 +1,119 @@
+# SPDX-FileCopyrightText: 2025 ETH Zurich and University of Bologna
+#
+# SPDX-License-Identifier: Apache-2.0
+"""
+Shared helpers used by the training / optimizer code-generation entry points
+(generateTrainingNetwork.py, testMVPTraining.py, testMVPOptimizer.py).
+
+These helpers read metadata and reference values out of inputs.npz / outputs.npz
+produced by the training ONNX exporter, and provide the singleton-pattern
+"scheduler" the Tiler expects when each node is handled independently.
+"""
+
+import os
+from typing import List, Optional
+
+import numpy as np
+import onnx_graphsurgeon as gs
+
+from Deeploy.Logging import DEFAULT_LOGGER as log
+
+# Graph input name marker identifying gradient accumulation buffers.
+_GRAD_ACC = "_grad.accumulation.buffer"
+
+
+def _load_reference_losses(train_dir: str) -> Optional[list]:
+    """Load reference loss values from outputs.npz.
+
+    Returns the list of per-mini-batch loss values if any key in
+    outputs.npz contains 'loss', otherwise None (with a warning).
+    """
+    outputs_path = os.path.join(train_dir, "outputs.npz")
+    if not os.path.exists(outputs_path):
+        log.warning(f"outputs.npz not found at {outputs_path} — loss comparison skipped")
+        return None
+
+    try:
+        outputs = np.load(outputs_path)
+    except Exception as e:
+        log.warning(f"Failed to load outputs.npz: {e} — loss comparison skipped")
+        return None
+
+    for key in outputs.files:
+        if 'loss' in key.lower():
+            vals = [float(v) for v in np.array(outputs[key]).flatten().tolist()]
+            log.info(f"Reference losses loaded from outputs.npz['{key}']: {vals}")
+            return vals
+
+    log.warning("No 'loss' key found in outputs.npz — loss comparison skipped")
+    return None
+
+
+def _infer_num_data_inputs(inputs_path: str) -> int:
+    """Auto-detect number of data inputs from inputs.npz.
+
+    Data inputs are the base arr_* entries that have per-mini-batch
+    variants (mb1_arr_*) in the npz — i.e. entries that actually change
+    across mini-batches.
+
+    Raises ValueError if no mb1 entries are found (single-mini-batch case)
+    where the data/weight boundary cannot be determined automatically.
+    """
+    inputs = np.load(inputs_path)
+    base_keys = sorted(k for k in inputs.files if not k.startswith('mb') and not k.startswith('meta_'))
+    count = sum(1 for k in base_keys if f'mb1_{k}' in inputs.files)
+    if count == 0:
+        raise ValueError("Cannot auto-detect num_data_inputs: inputs.npz has only one mini-batch "
+                         "(no mb1_arr_* entries found). Please pass --num-data-inputs explicitly.")
+    return count
+
+
+def _infer_total_mb(inputs_path: str) -> int:
+    """Count total mini-batches from inputs.npz.
+
+    New format: inputs.npz contains meta_n_batches (total training mini-batches)
+    and meta_data_size (number of unique samples stored; C harness cycles via modulo).
+
+    Legacy format: count 1 + number of unique mb* indices.
+    """
+    inputs = np.load(inputs_path)
+    if "meta_n_batches" in inputs.files:
+        return int(inputs["meta_n_batches"].flat[0])
+    mb_indices = set()
+    for key in inputs.files:
+        if key.startswith('mb'):
+            try:
+                idx = int(key.split('_')[0][2:])
+                mb_indices.add(idx)
+            except ValueError:
+                pass
+    return 1 + len(mb_indices)
+
+
+def _infer_data_size(inputs_path: str) -> int:
+    """Return the number of unique input samples stored in inputs.npz.
+
+    New format: reads meta_data_size.
+    Legacy format: same as _infer_total_mb (all batches were unique).
+    """
+    inputs = np.load(inputs_path)
+    if "meta_data_size" in inputs.files:
+        return int(inputs["meta_data_size"].flat[0])
+    return _infer_total_mb(inputs_path)
+
+
+def _infer_n_accum(inputs_path: str) -> int:
+    """Return the gradient accumulation step count stored in inputs.npz.
+
+    New format: reads meta_n_accum written by the exporter.
+    Legacy format: defaults to 1 (no gradient accumulation).
+    """
+    inputs = np.load(inputs_path)
+    if "meta_n_accum" in inputs.files:
+        return int(inputs["meta_n_accum"].flat[0])
+    return 1
+
+
+def _mockScheduler(graph: gs.Graph) -> List[List[gs.Node]]:
+    """Wrap every node in a singleton list for the Tiler pattern interface."""
+    return [[node] for node in graph.nodes]
